Standardize GPU device identifiers to int32, fix gosec G115/G109 warnings, and align CGO boundaries across all backends

+++ cmd/longbow/main.go
@@ -319,7 +319,7 @@ func run() error {
 		detectedBackend := gpu.DetectGPUBackend()
 
 		if detectedBackend == gpu.BackendCUDA || detectedBackend == gpu.BackendMetal {
-			vectorStore.SetGPUConfig(detectedBackend, cfg.GPUDeviceID)
+			vectorStore.SetGPUConfig(detectedBackend, int32(cfg.GPUDeviceID))
 			logger.Info().
 				Str("backend", detectedBackend.String()).
 				Bool("enabled", cfg.GPUEnabled).
